Error handling

+++ TomeOfHolding.BLL/CampaignService.cs
@@ -1,4 +1,5 @@
-﻿using TomeOfHolding.DAL;
+﻿using TomeOfHolding.BLL.Exceptions;
+using TomeOfHolding.DAL;
 using TomeOfHolding.Models;
 
 namespace TomeOfHolding.BLL {
@@ -10,11 +11,18 @@ namespace TomeOfHolding.BLL {
 		}
 
 		public async Task<List<Campaign>> GetCampaigns() {
-			return await _campaignRepo.GetCampaigns();
+			List<Campaign>? campaigns = await _campaignRepo.GetCampaigns();
+			if (campaigns == null || campaigns.Count == 0) {
+				throw new NotFoundException("No campaigns found.");
+			} else {
+				return campaigns;
+			}
+
 		}
 
 		public async Task<Campaign> GetCampaignById(int id) {
-			return await _campaignRepo.GetCampaignById(id);
+			Campaign campaign = await _campaignRepo.GetCampaignById(id);
+			return campaign ?? throw new NotFoundException("No campaign with that ID found.");
 		}
 
 		public async Task CreateCampaign(Campaign campaign) {
@@ -22,12 +30,16 @@ namespace TomeOfHolding.BLL {
 			await _campaignRepo.CreateCampaign(campaign);
 		}
 
-        public async Task DeleteCampaign(int id) {
-            await _campaignRepo.DeleteCampain(id);
-        }
+		public async Task DeleteCampaign(int id) {
+			Campaign? campaign = await _campaignRepo.GetCampaignById(id);
+			if (campaign == null) throw new NotFoundException("No campaign with that ID found."); 
+			await _campaignRepo.DeleteCampain(id);
+		}
 
-        public async Task UpdateCampaign(Campaign campaign) {
-            await _campaignRepo.UpdateCampain(campaign);
-        }
-    }
+		public async Task UpdateCampaign(Campaign campaign) {
+			Campaign? existingCampaign = await _campaignRepo.GetCampaignById(campaign.CampaignId);
+			if (existingCampaign == null) { throw new NotFoundException("No campaign with that ID found."); }
+			await _campaignRepo.UpdateCampain(campaign);
+		}
+	}
 }
